deka-package/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "deka"
version = "0.1.0"
authors = [
    {name = "Josue Godeme", email = "your.email@example.com"}
]
description = "A unified Python SDK for multiple translation providers"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["translation", "api", "google-translate", "deepl", "openai", "multilingual"]

dependencies = [
    "requests>=2.25.0",
    "aiohttp>=3.8.0",
    "pydantic>=1.8.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.18.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "build>=0.8.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://deka-website.vercel.app/"
Repository = "https://github.com/FrejusGdm/deka-project"
Documentation = "https://deka.readthedocs.io"
Issues = "https://github.com/FrejusGdm/deka-project/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/deka"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
    "/README.md",
    "/LICENSE",
]
